refactor: move the stamps to the correct place
