+++ src/api/receipts/mutations/index.ts
@@ -1,1 +1,2 @@
 export * from './useUploadFile';
+export * from './usePutAction';
