+++ src/api/receipts/queries/useGetExampleReceipts.ts
