+++ src/modules/receipts/components/ReceiptStatus/ReceiptStatus.tsx
@@ -1,6 +1,6 @@
 import { Text, TextProps } from '@chakra-ui/react';
 
-import { GetActionsResponse } from '../../api';
+import { GetActionsResponse } from '@/api';
 
 type ReceiptStatusProps = {
   status: GetActionsResponse['data'][0]['status'];
