+++ src/modules/receipts/components/Segment/Segment.tsx
@@ -1,6 +1,6 @@
 import { Badge } from '@chakra-ui/react';
 
-import { GetActionsResponse } from '../../api';
+import { GetActionsResponse } from '@/api';
 
 type SegmentProps = {
   status: GetActionsResponse['data'][0]['type'];
