+++ src/modules/receipts/pages/ReceiptsScreen/stores/useActionStore.ts
@@ -1,6 +1,6 @@
 import { create } from 'zustand';
 
-import { ActionResponse } from '@/modules/receipts/api';
+import { ActionResponse } from '@/api';
 
 type UseActionStore = {
   selectedAction: ActionResponse | null;
